Added item protection

+++ src/catalog/routes.py
@@ -3,6 +3,7 @@ import uuid
 from src.models import Item, Category
 from src.catalog.forms import ItemForm
 from src import db
+from flask_login import login_required, current_user
 
 bp = Blueprint('catalog', __name__)
 
@@ -22,28 +23,37 @@ def show_items_category(category_name):
 @bp.route('/items/<item_name>/', methods=['GET'])
 def view_item(item_name):
     item = Item.query.filter_by(name=item_name).one()
-    return render_template('catalog/view_item.html', item=item, title=item_name)
+    return render_template('catalog/view_item.html', item=item, title=item_name, current_user=current_user)
 
 @bp.route('/items/add/', methods=['GET','POST'])
+@login_required
 def add_item():
     form = ItemForm()
     if request.method == 'GET':
         return render_template('catalog/add_item.html', form=form, title='Add item')
     else:
+        
         item = Item(
             name = form.data['name'],
             description = form.data['description'],
-            category_id = form.data['category']
+            category_id = form.data['category'],
+            creator = current_user
             )
         db.session.add(item)
         db.session.commit()
         flash('Item added successfully!')
 
-        return(redirect(url_for('show_all_items')))
+        return(redirect(url_for('catalog.show_all_items')))
 
 @bp.route('/items/<item_name>/edit/', methods=['GET','POST'])
+@login_required
 def edit_item(item_name):
     item = Item.query.filter_by(name=item_name).one()
+
+    if current_user != item.creator:
+        flash("You are not authorized to edit this item, as you didn't create it!")
+        return redirect(url_for('catalog.view_item', item_name=item_name))
+
     form = ItemForm(name=item.name, description=item.description, category=item.category.name, title='Edit item')
 
     if request.method == 'GET':
@@ -56,11 +66,16 @@ def edit_item(item_name):
         db.session.commit()
         flash('Item edited successfully!')
 
-        return(redirect(url_for('show_all_items')))
+        return(redirect(url_for('catalog.show_all_items')))
 
 @bp.route('/items/<item_name>/delete/', methods=['GET','POST'])
+@login_required
 def delete_item(item_name):
     item = Item.query.filter_by(name=item_name).one()
+
+    if current_user != item.creator:
+        flash("You are not authorized to delete this item, as you didn't create it!")
+        return redirect(url_for('catalog.view_item', item_name=item_name))
 
     if request.method == 'GET':
         return render_template('catalog/delete_item.html', item=item, title='Delete item')
@@ -69,4 +84,4 @@ def delete_item(item_name):
         db.session.commit()
         flash('Item deleted successfully!')
 
-        return(redirect(url_for('show_all_items')))
+        return(redirect(url_for('catalog.show_all_items')))
